+++ src/core/client/views/hud/html/components/phone.js
@@ -40,7 +40,7 @@ const phone = Vue.component('phone', {
                 return;
             }
 
-            alt.emit('mouse:Focus', this.isActive);
+            alt.emit('mouse:Focus', this.isActive, 'isPhoneOpen');
         }
     },
     computed: {